improve(celery): Separate tasks from utils
