+++ open_crawler/celery_broker/tasks.py
@@ -1,5 +1,4 @@
 # Standard library imports
-import json
 import os
 import pathlib
 import shutil
@@ -7,58 +6,28 @@ from multiprocessing import Process, Manager
 
 # Third-party imports
 from minio import Minio
-from scrapy.crawler import CrawlerProcess
-from scrapy.utils.project import get_project_settings
-from scrapy.utils.python import without_none_values
 
 # Local imports
 import repositories
+from celery_broker.crawler_utils import start_crawler_process
 from celery_broker.main import celery_app
-from crawler.spiders.menesr import MenesrSpider
+from celery_broker.metadata_utils import metadata_task
+from celery_broker.utils import assume_content_type
 from models.crawl import CrawlModel
 from models.enums import MetadataType, ProcessStatus
 from models.metadata import MetadataTask
 from models.process import CrawlProcess
 from services.accessibility_best_practices_calculator import (
     LighthouseWrapper,
-    AccessibilityError,
-    BestPracticesError,
 )
-from services.carbon_calculator import CarbonCalculator, CarbonCalculatorError
+from services.carbon_calculator import CarbonCalculator
 from services.crawler_logger import logger, set_file
 from services.responsiveness_calculator import (
     ResponsivenessCalculator,
-    ResponsivenessCalculatorError,
 )
 from services.technologies_calculator import (
     TechnologiesCalculator,
-    TechnologiesError,
 )
-
-
-def update_crawl_status(crawl: CrawlModel, status: ProcessStatus):
-    crawl.update_status(status=status)
-    repositories.crawls.update(crawl)
-
-
-def init_crawler_settings(crawl_process: CrawlProcess):
-    settings = get_project_settings()
-    custom_settings = without_none_values(
-        {
-            "DEPTH_LIMIT": crawl_process.config.parameters.depth,
-            "CLOSESPIDER_PAGECOUNT": crawl_process.config.parameters.limit,
-            "CUSTOM_HEADERS": crawl_process.config.headers or {},
-        }
-    )
-    settings.update(custom_settings)
-    return settings
-
-
-def start_crawler_process(crawl_process: CrawlProcess, results: dict):
-    process = CrawlerProcess(settings=init_crawler_settings(crawl_process))
-    process.crawl(MenesrSpider, crawl_process=crawl_process)
-    process.start()
-    results["metadata"] = dict(crawl_process.metadata.items())
 
 
 @celery_app.task(bind=True, name="crawl")
@@ -94,100 +63,6 @@ def start_crawl_process(self, crawl: CrawlModel) -> CrawlProcess:
     )
     logger.debug("Html crawl ended!")
     return crawl_process
-
-
-def handle_metadata_result(
-    task: MetadataTask,
-    crawl_process: CrawlProcess,
-    result: dict,
-    metadata_type: MetadataType,
-):
-    if not result:
-        task.update(status=ProcessStatus.ERROR)
-        repositories.crawls.update_task(
-            crawl_id=crawl_process.id,
-            task_name=metadata_type,
-            task=task,
-        )
-        set_file(crawl_process.id)
-        logger.error(f"{metadata_type} failed.")
-        return
-    store_metadata_result(crawl_process, result, metadata_type)
-    if task.status == ProcessStatus.STARTED:
-        task.update(status=ProcessStatus.SUCCESS)
-        repositories.crawls.update_task(
-            crawl_id=crawl_process.id,
-            task_name=metadata_type,
-            task=task,
-        )
-        logger.debug(f"{metadata_type} ended!")
-    return result
-
-
-def store_metadata_result(
-    crawl_process: CrawlProcess, result: dict, metadata_type: MetadataType
-):
-    base_file_path = f"{os.environ['LOCAL_FILES_PATH']}/{crawl_process.id}"
-    file_path = pathlib.Path(
-        f"{base_file_path}/{os.environ['METADATA_FOLDER_NAME']}/{metadata_type}.json"
-    )
-    file_path.parent.mkdir(exist_ok=True, parents=True)
-    file_path.write_text(json.dumps(result, indent=4))
-
-
-def metadata_task(
-    task: MetadataTask,
-    crawl_process: CrawlProcess,
-    metadata_type: MetadataType,
-    calculator,
-    method_name: str,
-):
-    calc_method = getattr(calculator, method_name)
-    result = {}
-    task.update(status=ProcessStatus.STARTED)
-    set_file(crawl_process.id)
-    logger.debug(f"{metadata_type} started!")
-    repositories.crawls.update_task(
-        crawl_id=crawl_process.id,
-        task_name=metadata_type,
-        task=task,
-    )
-    if metadata_process := crawl_process.metadata.get(metadata_type):
-        for url in metadata_process.urls:
-            try:
-                data = calc_method(url)
-                result[url] = data
-            except (
-                AccessibilityError,
-                BestPracticesError,
-                TechnologiesError,
-                ResponsivenessCalculatorError,
-                CarbonCalculatorError,
-            ) as e:
-                logger.warning(
-                    f"An error occured for url {url} during {metadata_type} process. {e}"
-                )
-                if task.status != ProcessStatus.PARTIAL_ERROR:
-                    task.update(status=ProcessStatus.PARTIAL_ERROR)
-                    repositories.crawls.update_task(
-                        crawl_id=crawl_process.id,
-                        task_name=metadata_type,
-                        task=task,
-                    )
-                continue
-            except Exception as e:
-                logger.error(
-                    f"An unknown error occured for url {url} during {metadata_type} process. {e}"
-                )
-                if task.status != ProcessStatus.PARTIAL_ERROR:
-                    task.update(status=ProcessStatus.PARTIAL_ERROR)
-                    repositories.crawls.update_task(
-                        crawl_id=crawl_process.id,
-                        task_name=metadata_type,
-                        task=task,
-                    )
-                continue
-    return handle_metadata_result(task, crawl_process, result, metadata_type)
 
 
 @celery_app.task(bind=True, name="get_accessibility")
@@ -303,15 +178,6 @@ def upload_html(self, crawl: CrawlModel):
         website_id=crawl.website_id,
         crawl=repositories.crawls.get(crawl_id=crawl.id).model_dump(),
     )
-
-
-def assume_content_type(file_path: str) -> str:
-    # sourcery skip: assign-if-exp, reintroduce-else
-    if file_path.endswith("html"):
-        return "text/html"
-    if file_path.endswith("json"):
-        return "application/json"
-    return ""
 
 
 METADATA_TASK_REGISTRY = {
